requirements.txt
random
scipy.stats
pandas
streamlit